integer

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ long int li;
 long long int lli;
 float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of an int: %lu bytes(s)\n",(unsigned long)sizeof(i));
-printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(li));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(lli));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %d byte(s)\n", (int)sizeof(c));
+printf("Size of an int: %d bytes(s)\n", (int)sizeof(i));
+printf("Size of a long int: %d bytes(s)\n", (int)sizeof(li));
+printf("Size of a long long int: %d byte(s)\n", (int)sizeof(lli));
+printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 
 return (0);
 }
